회원가입/수정API  카테고리리스트 필터추가

+++ src/client-modules/member/member.service.ts
@@ -43,7 +43,9 @@ export class MemberService {
 
         await this.memberRepository.updateMemberNickname(newMemberId, nickName);
 
-        const categories = [signInDto.memberPickCategory1, signInDto.memberPickCategory2, signInDto.memberPickCategory3, signInDto.memberPickCategory4, signInDto.memberPickCategory5];
+        const categories = [signInDto.memberPickCategory1, signInDto.memberPickCategory2, signInDto.memberPickCategory3, signInDto.memberPickCategory4, signInDto.memberPickCategory5].filter(
+          (category) => category !== undefined,
+        );
         categories.forEach((category) => {
           if (category) {
             pickedDCList.push({
@@ -52,7 +54,10 @@ export class MemberService {
             });
           }
         });
-        await this.memberRepository.insertPickCategoryList(pickedDCList);
+
+        if (categories.length > 0) {
+          await this.memberRepository.insertPickCategoryList(pickedDCList);
+        }
         return await this.memberRepository.findSnsId(newMemberId);
       } else {
         throw new BadRequestException('중복정보', {
@@ -105,8 +110,6 @@ export class MemberService {
       const usersReviewCount = await this.memberRepository.countReview(memberIdDto);
       const usersPoint = await this.memberRepository.findTotalPointOne(memberIdDto);
       const usersTotalPoint = usersPoint[0] ? usersPoint[0].totalPoint : 0;
-      console.log('usersReviewCount::', usersReviewCount);
-      console.log('usersTotalPoint::', usersTotalPoint);
       return {
         nickName: nickName.nickName,
         usersReviewCount,
@@ -190,8 +193,9 @@ export class MemberService {
       await this.memberRepository.saveMember(memberUpdateDto);
 
       const pickDessertList = [];
-      const categories = [memberUpdateDto.memberPickCategory1, memberUpdateDto.memberPickCategory2, memberUpdateDto.memberPickCategory3, memberUpdateDto.memberPickCategory4, memberUpdateDto.memberPickCategory5];
-
+      const categories = [memberUpdateDto.memberPickCategory1, memberUpdateDto.memberPickCategory2, memberUpdateDto.memberPickCategory3, memberUpdateDto.memberPickCategory4, memberUpdateDto.memberPickCategory5].filter(
+        (category) => category !== undefined,
+      );
       categories.forEach((category) => {
         if (category) {
           pickDessertList.push({
@@ -200,9 +204,10 @@ export class MemberService {
           });
         }
       });
-
-      await this.memberRepository.deletePickCategoryList(memberUpdateDto);
-      await this.memberRepository.insertPickCategoryList(pickDessertList);
+      if (categories.length > 0) {
+        await this.memberRepository.deletePickCategoryList(memberUpdateDto);
+        await this.memberRepository.insertPickCategoryList(pickDessertList);
+      }
     } catch (error) {
       throw error;
     }
